+++ Bjorn/Cards/SafetyProtocolCard.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using Nanoray.PluginManager;
 using Nickel;
@@ -9,8 +8,6 @@ namespace Shockah.Bjorn;
 
 public sealed class SafetyProtocolCard : Card, IRegisterable, IHasCustomCardTraits
 {
-	private int GetDataReentry;
-	
 	public static void Register(IPluginPackage<IModManifest> package, IModHelper helper)
 	{
 		var entry = helper.Content.Cards.RegisterCard(MethodBase.GetCurrentMethod()!.DeclaringType!.Name, new()
@@ -25,29 +22,6 @@ public sealed class SafetyProtocolCard : Card, IRegisterable, IHasCustomCardTrai
 			Art = helper.Content.Sprites.RegisterSpriteOrDefault(package.PackageRoot.GetRelativeFile("assets/Cards/SafetyProtocol.png"), StableSpr.cards_Inverter).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "SafetyProtocol", "name"]).Localize,
 		});
-
-		helper.Content.Cards.OnGetDynamicInnateCardTraitOverrides += (_, e) =>
-		{
-			if (e.Card is not SafetyProtocolCard card)
-				return;
-
-			card.GetDataReentry++;
-			try
-			{
-				var unplayable = card.GetDataReentry == 1 && e.State.route is Combat { routeOverride: null } combat && (
-					card.flipped
-						? combat.energy < card.GetCurrentCost(MG.inst.g?.state ?? DB.fakeState) + 1
-						: combat.hand.All(someCard => !someCard.IsAnalyzable(MG.inst.g?.state ?? DB.fakeState, combat))
-				);
-				
-				if (unplayable)
-					e.SetOverride(ModEntry.Instance.Helper.Content.Cards.UnplayableCardTrait, true);
-			}
-			finally
-			{
-				card.GetDataReentry--;
-			}
-		};
 		
 		ModEntry.Instance.KokoroApi.Finite.SetBaseFiniteUses(entry.UniqueName, Upgrade.None, 2);
 		ModEntry.Instance.KokoroApi.Finite.SetBaseFiniteUses(entry.UniqueName, Upgrade.B, 2);